storage: refine file cache related code
Refine file cache related code by:
1) merge cache_entry.rs into filecache/mod.rs
2) rename BlobCacheConfig to FileCacheConfig
